`requestTask` now accepts a decoder instead of a resolver.
The previous version of `requestTask` left the resolver up to the user and let them determine the type of response (json / string / etc). The response type from StandardAPI will always be JSON. This change simplifies the creation fo requests by allowing the user to pass in a decoder instead of crafting his or her own resolver.